backend/pyproject.toml
[tool.poetry]
name = "yeetchess-backend"
version = "0.1.0"
description = "YeetChess FastAPI Backend"
authors = ["YeetChess <dev@yeetchess.local>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0"
asyncpg = "^0.29.0"
pydantic = "^2.5"
pydantic-settings = "^2.1"
pydantic-extra-types = "^2.0"
email-validator = "^2.1"
bcrypt = "^4.1.1"
pyjwt = "^2.8.1"
python-multipart = "^0.0.6"
redis = "^5.0.1"
python-chess = "^1.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
ruff = "^0.1.8"
black = "^23.12.0"
mypy = "^1.7.1"
httpx = "^0.25.0"
aiosqlite = "^0.22.1"
pytest-cov = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
